fix a compile error and remove warnings.

+++ library/androidndkgif/jni/GCTGifEncoder.cpp
@@ -114,7 +114,7 @@ void GCTGifEncoder::buildColorTable(Cube cubes[256]) {
 	for (std::vector<FrameInfo*>::iterator i = images.begin(); i != images.end(); ++i, ++idx) {
 		memcpy(allPixels + width * height * idx, (*i)->pixels, width * height * sizeof(allPixels[0]));
 	}
-	
+
 	computeColorTable(allPixels, cubes, pixelNum);
 
 	delete[] allPixels;
@@ -235,7 +235,7 @@ bool GCTGifEncoder::writeGraphicControlExt(uint16_t delay)
 
 	uint8_t packed = (disposalMethod << 2) | (userInputFlag << 1) | transparencyFlag;
 	//                                                     size, packed, delay(2), transIndex, terminator
-	const uint8_t graphicControlExt[] = {0x21, 0xF9, 0x04, packed, delay & 0xFF, (delay >> 8), 0xFF, 0x00};
+	const uint8_t graphicControlExt[] = {0x21, 0xF9, 0x04, packed, (uint8_t)(delay & 0xFF), (uint8_t)(delay >> 8), 0xFF, 0x00};
 	fwrite(graphicControlExt, sizeof(graphicControlExt), 1, fp);
 	return true;
 }
